rename directory to integration to fit convention
